requirements.txt
flask
ultralytics
werkzeug
python-dotenv
opencv-python
